Refactoring metrics into standalone modules
